requirements.txt


asgiref-3.2.10 
django-3.1 
pytz-2020.1 
sqlparse-0.3.1
xhtml2pdf==0.2.4
django-decouple-2.1